+++ demisto_sdk/commands/lint/resources/pylint_plugins/xsoar_level_checker.py
@@ -33,7 +33,9 @@ class XsoarChecker(BaseChecker):
     def __init__(self, linter=None):
         super(XsoarChecker, self).__init__(linter)
         self.is_script = True if os.getenv('is_script') == 'True' else False
-        self.common_args_params = ['args', 'dargs', 'arguments', 'd_args', 'data_args', 'params', 'PARAMS', 'integration_parameters']
+        self.common_args_params = ['args', 'dargs', 'arguments', 'd_args', 'data_args', 'params', 'PARAMS',
+                                   'integration_parameters']
+        self.common_params_names = ['params', 'PARAMS', 'integration_parameters']
 
     def visit_functiondef(self, node):
         self._type_annotations_checker(node)
@@ -74,14 +76,18 @@ class XsoarChecker(BaseChecker):
 
     def _direct_access_dict_checker(self, node):
         try:
-            # check if the direct usage is not an assign to dict like a['a'] = b
-            if not isinstance(node.parent, astroid.Assign) and isinstance(node.value, astroid.Name) and \
-                    node.value.name in self.common_args_params:
-                for infered in node.value.infer():
-                    if isinstance(infered, astroid.Dict):
-                        self.add_message("direct-access-dict-exist", node=node)
+            # for demisto.args()[] implementation or for demisto.params()[]
+            if node.value.func.expr.name == 'demisto' and node.value.func.attrname == 'args':
+                self.add_message("direct-access-dict-exist", node=node)
+            elif node.value.func.expr.name == 'demisto' and node.value.func.attrname == 'params':
+                self.add_message("direct-access-dict-exist", node=node)
         except Exception:
-            pass
+            try:
+                # for args[] implementation
+                if node.value.name in self.common_args_params or node.value.name in self.common_params_names:
+                    self.add_message("direct-access-dict-exist", node=node)
+            except Exception:
+                pass
 
     def _inner_search_return_error(self, node):
         try: