+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * SDK repository is now mypy check_untyped_defs complaint.
 * The lint command will now ignore the unsubscriptable-object (E1136) pylint error in dockers based on python 3.9 - this will be removed once a new pylint version is released.
 * Added an option for **format** to run on a whole pack.
+* Added new validation of direct access to args/params dicts to `XSOAR-linter`.
 
 # 1.2.13
 * Added new validation of indicators usage in CommandResults to `XSOAR-linter`.